Clean up code for portability and server initialization
* Practically move all non-commented code from main.cpp
* Create Client.cpp/h class and move logic out of main.cpp
* Create default constructor for Renderer
* Fix definition of default:stone
* Change GameInstance:update() to use a double instead of a GLfloat

+++ zeus/game/GameInstance.cpp
@@ -75,7 +75,7 @@ void GameInstance::initialize(Renderer* renderer) {
     guiEntities.push_back(crosshair);
 }
 
-void GameInstance::update(GLfloat deltaTime, double fps) {
+void GameInstance::update(double deltaTime, double fps) {
     renderer->update();
 
     auto camera = renderer->getCamera();
